refactor(all): move the observers together
