+++ src/runtime/binding/resources/update-trigger-binding-behavior.ts
@@ -2,9 +2,9 @@ import { EventSubscriber, IEventSubscriber } from '../event-manager';
 import { IObserverLocator } from '../observer-locator';
 import { Binding, BindingMode } from '../binding';
 import { IScope } from '../binding-context';
-import { ValueAttributeObserver } from '../element-observation';
-import { CheckedObserver } from '../checked-observer';
-import { SelectValueObserver } from '../select-value-observer';
+import { ValueAttributeObserver } from '../observers/element-observation';
+import { CheckedObserver } from '../observers/checked-observer';
+import { SelectValueObserver } from '../observers/select-value-observer';
 import { Reporter } from '../../../kernel/reporter';
 import { bindingBehavior } from '../binding-behavior';
 import { inject } from '../../../kernel/di';
