+++ test/runtime/binding/primitive-observer.spec.ts
@@ -1,4 +1,4 @@
-import { PrimitiveObserver } from '../../../src/runtime/binding/property-observation';
+import { PrimitiveObserver } from '../../../src/runtime/binding/observers/property-observation';
 import { createObserverLocator } from './shared';
 import { expect } from 'chai';
